arregle las clases en los 2 proyectos, y ya funciona el abrir caja

+++ Proyecto2/Proyecto2.ClienteWeb/Models/Abrir_Caja.cs
@@ -8,7 +8,7 @@ namespace Proyecto2.ClienteWeb.Models
     public class Abrir_Caja
     {
         public DateTime FechaHora { get; set; }
-        public int Monto { get; set; }
+        public double Monto { get; set; }
         public int Usuario { get; set; }
         public int Caja { get; set; }
     }
