+++ Proyecto2/Proyecto2.ClienteWeb/Models/Cuenta.cs
@@ -10,8 +10,8 @@ namespace Proyecto2.ClienteWeb.Models
         public int cuenta { get; set; }
         public int NoFactura { get; set; }
         public DateTime FechaEmision { get; set; }
-        public int TotalFactura { get; set; }
-        public int SaldoFactura { get; set; }
+        public double TotalFactura { get; set; }
+        public double SaldoFactura { get; set; }
         public int Cliente { get; set; }
         public int EstadoCuenta { get; set; }
     }
