+++ Proyecto2/Proyecto2.ClienteWeb/Models/Venta.cs
@@ -10,8 +10,8 @@ namespace Proyecto2.ClienteWeb.Models
         public int venta { get; set; }
         public int NIT { get; set; }
         public string NombreCliente { get; set; }
-        public int IVA { get; set; }
-        public int MontoPagado { get; set; }
+        public double IVA { get; set; }
+        public double MontoPagado { get; set; }
         public int Usuario { get; set; }
         public int Caja { get; set; }
         public int Factura { get; set; }
