+++ Proyecto2/Proyecto2.WebApi/Models/Caja.cs
@@ -8,10 +8,10 @@ namespace Proyecto2.WebApi.Models
     public class Caja
     {
         public int caja { get; set; }
-        public int Monto { get; set; }
+        public double Monto { get; set; }
         public int Estado { get; set; }
 
-        public Caja(int caja,int monto,int estado)
+        public Caja(int caja,double monto,int estado)
         {
             this.caja = caja;
             this.Monto = monto;
