+++ Proyecto2/Proyecto2.WebApi/Models/Transaccion_Cuenta.cs
@@ -8,11 +8,11 @@ namespace Proyecto2.WebApi.Models
     public class Transaccion_Cuenta
     {
         public int TransaccionCuenta { get; set; }
-        public int Monto { get; set; }
+        public double Monto { get; set; }
         public int Cuenta { get; set; }
         public int Tipo_Transaccion { get; set; }
 
-        public Transaccion_Cuenta(int TransaccionCuenta,int monto,int cuenta,int tipoTransaccion)
+        public Transaccion_Cuenta(int TransaccionCuenta,double monto,int cuenta,int tipoTransaccion)
         {
             this.TransaccionCuenta = TransaccionCuenta;
             this.Monto = monto;
